final testing approved

+++ src/code/GUI/Deposit.java
@@ -10,7 +10,7 @@ public class Deposit extends Transaction
    private Euro amount; // amount to deposit
    private Keypad keypad; // reference to keypad
    private DepositSlot depositSlot; // reference to deposit slot
-   private final static int CANCELED = 0; // constant for cancel option
+   private final static double CANCELED = 0.0; // constant for cancel option
 
    // Deposit constructor
    public Deposit( int userAccountNumber, Screen atmScreen,
